+++ tests/test_logging_slack.py
@@ -1,10 +1,11 @@
 import logging
+import sys
 from logging import LogRecord
 
 import pytest
 from mock import patch, MagicMock
 
-from pnp.logging import SlackHandler, DEFAULT_USERNAME, DEFAULT_EMOJI
+from pnp.logging import SlackHandler
 
 
 @patch('pnp.logging.slacker.Slacker')
@@ -29,12 +30,52 @@ def test_logger_emit_for_smoke(slacker_mock):
     slacker_mock.return_value.chat.post_message.assert_called_once_with(
         text=None,
         channel='#pytest',
-        username=DEFAULT_USERNAME,
+        username=SlackHandler.DEFAULT_USERNAME,
         icon_url=None,
-        icon_emoji=DEFAULT_EMOJI,
+        icon_emoji=SlackHandler.DEFAULT_EMOJI,
         attachments=[{
             'color': 'danger',
             'fields': [{'title': 'LogRecord from pytest', 'short': False}]
+        }]
+    )
+
+
+@patch('pnp.logging.slacker.Slacker')
+def test_logger_emit_with_trace(slacker_mock):
+    slacker_mock.return_value = MagicMock()
+
+    dut = SlackHandler(
+        api_key='doesnt_matter',
+        channel='pytest',
+        fire_and_forget=False
+    )
+    dut.MAX_ATTACHMENT_CHARS = 46
+    try:
+        raise Exception("EXCEPTION RAISED ON PURPOSE!")
+    except Exception:
+        dut.emit(LogRecord(
+            name='pytest',
+            level=logging.ERROR,
+            pathname='doesnt_matter',
+            lineno=42,
+            msg='LogRecord from pytest',
+            args=None,
+            exc_info=sys.exc_info()
+        ))
+
+    slacker_mock.return_value.chat.post_message.assert_called_once_with(
+        text=None,
+        channel='#pytest',
+        username=SlackHandler.DEFAULT_USERNAME,
+        icon_url=None,
+        icon_emoji=SlackHandler.DEFAULT_EMOJI,
+        attachments=[{
+            'color': 'danger',
+            'fields': [{
+                'title': 'LogRecord from pytest',
+                'short': False,
+                'value': '```Exception: EXCEPTION RAISED ON PURPOSE!\n```'
+            }]
         }]
     )
 
@@ -74,9 +115,9 @@ def test_ping_users_existing_user(slacker_mock):
     slacker_mock.return_value.chat.post_message.assert_called_once_with(
         text="<@42> ",
         channel='#pytest',
-        username=DEFAULT_USERNAME,
+        username=SlackHandler.DEFAULT_USERNAME,
         icon_url=None,
-        icon_emoji=DEFAULT_EMOJI,
+        icon_emoji=SlackHandler.DEFAULT_EMOJI,
         attachments=[{
             'color': 'danger',
             'fields': [{'title': 'LogRecord from pytest', 'short': False}]